fixed issue with float conditionning not displayed properly

+++ BackEnd/model/Article.cs
@@ -179,8 +179,8 @@ namespace BackEnd.model
                 NotifyPropertyChanged();
             }
         }
-        private int _condi;
-        public int condi
+        private float _condi;
+        public float condi
         {
             get => _condi;
             set
